Should have deltaMs as a double for better performance with speed modulations

+++ src/heronarts/lx/Tempo.java
@@ -139,7 +139,7 @@ public class Tempo {
         }
     }
     
-    void run(int deltaMs) {
+    void run(double deltaMs) {
         this.click.run(deltaMs);
         if (this.beat()) {
             this.ramp.trigger();
